fix: address all 11 code review findings from PR #23
Blockers (high):
- OAuth redirect flow: add useEffect on user in LoginForm so mobile/popup-blocked
  users are navigated to their destination after signInWithRedirect completes
- Persist ?from= param to sessionStorage before redirect so destination survives
  the Firebase OAuth round-trip
- Cart localStorage restore: guard JSON.parse result with Array.isArray to prevent
  TypeError silently destroying the cart on corrupt/legacy storage shape

Medium:
- CartNotification: capture motion key in shownKey state so it stays stable during
  exit animation (was flipping number→string when clearLastAdded fired, hard-cutting)
- CartContext addItem: read existing quantity from items before setItems so toast
  shows correct qty on re-adds (was always showing 1)
- AUTH_COOKIE extracted to src/lib/constants.ts — shared by proxy.ts and AuthContext
- resolveDestination: block backslash open-redirect (/\evil.com) alongside //
- Price formatting: new formatPrice() in src/lib/format.ts (ZAR, en-ZA locale)
  replaces hardcoded $ symbol in CartSidebar, CartNotification, CheckoutContent
  and aligns line-item and total rounding

Low (CLAUDE.md):
- Auth pages split into server page.tsx (exports metadata) + use client *Content.tsx
- CartNotification easing: [0.16,1,0.3,1] → ease.cinematic from src/lib/motion.ts
