+++ src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const AUTH_COOKIE = "hb-auth-token";
